requirements.txt
# 591 Taiwan Apartment Scraper - Dependencies

# Browser automation
playwright>=1.40.0

# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0

# Google Sheets (optional - only needed for Sheets export)
google-api-python-client>=2.100.0
google-auth>=2.23.0
